+++ include/CKScene.h
@@ -215,7 +215,7 @@ public:
     virtual void PreDelete();
     virtual void CheckPostDeletion();
 
-    virtual int GetMemoryOccupation();
+    virtual size_t GetMemoryOccupation();
     virtual int IsObjectUsed(CKObject *o, CK_CLASSID cid);
 
     //--------------------------------------------
