+++ src/CK2_3D/CKLight.cpp
@@ -678,7 +678,7 @@ Remarks:
 
 Implementation based on decompilation at 0x1001b36b.
 *************************************************/
-int RCKLight::GetMemoryOccupation() {
+size_t RCKLight::GetMemoryOccupation() {
     return RCK3dEntity::GetMemoryOccupation() + (sizeof(RCKLight) - sizeof(RCK3dEntity));
 }
